добавил компоненты: Products, Slider, WalletConnectionButton
