+++ src/pages/Homepage/components/products/products.module.scss
@@ -1,0 +1,17 @@
+.catalog {
+    padding-inline: 15px;
+
+    display: grid;
+    grid-template-columns: repeat(2, calc(50% - 10px));
+    column-gap: 10px;
+    row-gap: 20px;
+    margin-top: 6px;
+}
+
+.catalogTitle {
+    margin-left: 15px;
+    margin-top: 30px;
+    font-size: 14px;
+    font-family: "H_neue";
+    line-height: 16px;
+}
