+++ src/pages/Homepage/components/slider/slides.module.scss
@@ -1,0 +1,34 @@
+.banner {
+    padding-inline: 15px;
+}
+
+.bannerItem {
+    border-radius: 12px;
+
+    img {
+        height: 100%;
+        width: 100%;
+        object-fit: cover;
+    }
+}
+
+.dots {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 3px;
+    margin-top: 4px;
+}
+
+.dot {
+    border-radius: 10px;
+    width: 6px;
+    height: 6px;
+    background-color: #c0c0c0;
+    transition: 0.2s;
+}
+
+.activeDot {
+    background-color: var(--blue-color);
+    width: 14px;
+}
