+++ src/pages/Homepage/homepage.module.scss
@@ -1,109 +1,40 @@
 .wrapper {
-	padding-top: 15px;
-	position: relative;
-	padding-bottom: 170px;
+    padding-top: 15px;
+    position: relative;
+    padding-bottom: 170px;
 
-	.banner {
-		padding-inline: 15px;
-	}
+    .categories {
+        padding-inline: 15px;
+        display: flex;
+        margin-top: 8px;
+        column-gap: 7px;
+        justify-content: space-between;
 
-	.bannerItem {
-		border-radius: 12px;
+        .catImageBlock {
+            background-color: #06a5ff;
+            width: 100%;
+            width: 100%;
+            aspect-ratio: 1;
+            border-radius: 12px;
+            max-width: 76px;
+            cursor: pointer;
+            display: flex;
+            align-items: flex-end;
 
-		img {
-			height: 100%;
-			width: 100%;
-			object-fit: cover;
-		}
-	}
+            img {
+                width: 100%;
+                height: 100%;
+                object-fit: cover;
+                object-position: bottom;
+            }
+        }
 
-	.bannerTitle {
-		font-family: 'H_neue';
-		color: white;
-		font-size: 24px;
-		line-height: 26px;
-		font-weight: 500;
-	}
-
-	.bannerDesc {
-		color: white;
-		font-size: 14px;
-		font-weight: 400;
-		font-family: 'H_neue';
-		margin-top: 10px;
-	}
+        .catText {
+            margin-top: 3px;
+            font-size: 11px;
+            font-weight: 500;
+            line-height: 11px;
+            text-align: center;
+        }
+    }
 }
-
-.categories {
-	padding-inline: 15px;
-	display: flex;
-	margin-top: 8px;
-	column-gap: 7px;
-	justify-content: space-between;
-
-	.catImageBlock {
-		background-color: #06a5ff;
-		width: 100%;
-		width: 100%;
-		aspect-ratio: 1;
-		border-radius: 12px;
-		max-width: 76px;
-		cursor: pointer;
-		display: flex;
-		align-items: flex-end;
-
-		img {
-			width: 100%;
-			height: 100%;
-			object-fit: cover;
-			object-position: bottom;
-		}
-	}
-
-	.catText {
-		margin-top: 3px;
-		font-size: 11px;
-		font-weight: 500;
-		line-height: 11px;
-		text-align: center;
-	}
-}
-
-.catalog {
-	padding-inline: 15px;
-
-	display: grid;
-	grid-template-columns: repeat(2, calc(50% - 10px));
-	column-gap: 10px;
-	row-gap: 20px;
-	margin-top: 6px;
-}
-
-.dots {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	gap: 3px;
-	margin-top: 4px;
-}
-
-.dot {
-	border-radius: 10px;
-	width: 6px;
-	height: 6px;
-	background-color: #c0c0c0;
-	transition: .2s;
-}
-
-.activeDot {
-	background-color: var(--blue-color);
-	width: 14px;
-}
-
- .catalogTitle {
-	margin-left: 15px;
-	margin-top: 30px;
-	font-size: 14px;
-	font-family: 'H_neue';
-	line-height: 16px;
- }
